Prototype of a prepend-message signal
erlang:send/3 support for 'prepend' option when receiver is identified
by a local PID.

When the 'prepend' option is used, the message is sent using a prepend-
message signal instead of using an ordinary message signal. When this
signal is received at the destination, the message is prepended to the
receivers message queue instead of being appended to the message queue as
done when an ordinary message signal is received.